Change test structure
